added demo cmake and moved monopod.cpp/hpp into the core component. This created dependency loop which needs fixing.
