+++ src/monopod_drivers/demos/demo_leg.cpp
@@ -6,8 +6,8 @@
 
 #include <math.h>
 #include <monopod_sdk/blmc_drivers/devices/analog_sensor.hpp>
-#include <monopod_sdk/blmc_drivers/devices/leg.hpp>
 #include <monopod_sdk/blmc_drivers/devices/motor.hpp>
+#include <monopod_sdk/monopod_drivers/leg.hpp>
 #include "real_time_tools/spinner.hpp"
 #include "real_time_tools/timer.hpp"
