+++ src/monopod_drivers/demos/sine_position_control.cpp
@@ -129,4 +129,4 @@ void SinePositionControl::stop_loop()
     rt_printf("dumped the trajectory");
 }
 
-}  // namespace blmc_drivers
+}  // namespace monopod_drivers
